read_bytes(): suppress PyCharm "Unbound local variable" inspection
PyCharm is right from the syntactical point of view, but the current
logic guarantees that if `is_satisfiable` is:

  - `False`, `num_bytes_available` is set (and can be used in the error
  message);
  - `True`, `r` is set, so it's safe to return.

I could "bind" the local variable even before doing the actual stuff
inside `if`s, but I exactly don't want to accidentally return the value
set that way. If the logic is screwed up like that, it's better to get a
`NameError` (so we quickly get notified about that) than return a random
value.