Move version definition to __init__.py
Adjusted the setup.cfg file to reflect changes in version attribute location. This was to standardize the way the version number is handled and stored. Added __init__.py file for storing the version number, as it's a more standard practice. This will also avoid redundancy and potential inconsistencies in maintaining the version number at multiple places.