read_bytes_term(): speed up by avoiding unnecessary memory copying
As noted in the previous commit, `bytes` is immutable, and thus
concatenating two `bytes` objects (even using the `+=` operation) incurs
a memory copy. This causes an unnecessary slowdown, because in
particular here in the `read_bytes_term` method, the contents of the
`bytes` object stored in `r` must be copied when appending each byte.

The updated implementation uses mutable `bytearray` instead of immutable
`bytes` to collect the bytes to be returned. To make this change
completely transparent to user applications, it is converted to `bytes`
before being returned using the `bytes()` constructor (this also incurs
a memory copy, but only once at the end, not for each byte added).

I've done some simple benchmarks
(https://gist.github.com/generalmimon/41dcb2bba48cb47c2f30ce9f086cbbf2)
to see how much this change helps. On Python 3.12.4 under Ubuntu
22.04.4, it seems to be ~1.5x faster if the terminator occurs after 1
KiB, ~2.1x faster for the terminator after 4 KiB and ~3.4x faster for
the terminator after 16 KiB.

BTW, you can use `ltrace -e memcpy python3 ...` (redirect it to a file
though, it will be several MiBs) to check that the `memcpy` function is
called for each length from 1 up to the final length of the output
variable `r`, so the number of `memcpy` calls grows linearly. The new
version apparently calls `memcpy` only a logarithmic number of times, as
expected (needed for the reallocation of `bytearray`).