Configure Ruff: enable all rules except selected ones
Although selecting `ALL` rules is not exactly recommended, see
<https://docs.astral.sh/ruff/linter/#rule-selection>:

> - Use `ALL` with discretion. Enabling `ALL` will implicitly enable new
>   rules whenever you upgrade.

... but I'd argue that's not a bad thing. In my opinion, ignoring any
lint rule should be a deliberate decision, not an implicit one. How can
you be sure that no new rules are useful? It's perfectly fine to
disable rules that we don't find useful, but I believe we should give
them a chance first.